remove all javascript file and move login and register in user folder
