+++ client/angular.json
@@ -128,5 +128,8 @@
         }
       }
     }
+  },
+  "cli": {
+    "analytics": "23f143a2-e558-4efb-82ee-121186b71811"
   }
 }
