--- client/src/polyfills.js
@@ -1,4 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-require("zone.js");
-//# sourceMappingURL=polyfills.js.map
